adding openmp utils
